chore: restore deleted reducers
