+++ src/app/[locale]/games/page.tsx
@@ -1,17 +1,28 @@
+"use client";
+
 // Hooks
-import {useTranslations} from 'next-intl';
+import { useSearchParams } from 'next/navigation'
+import dynamic from 'next/dynamic'
 
 // Client components
 import GalleryMode from './_client/GalleryMode';
+const GamesGalleryGrid = dynamic(() => import('./_client/GamesGalleryGrid'))
+const GamesGalleryList = dynamic(() => import('./_client/GamesGalleryList'))
+
+type ViewType = "GRID" | "LIST";
 
 // The gallery component
 export default function GamesPage() {
 
-    const t = useTranslations('gamesLibrary.tabs');
+    const searchParams = useSearchParams()
+    const viewMode : ViewType = searchParams.get("mode") === "LIST" ? "LIST" : "GRID";
 
     return (
         <>
-            <GalleryMode gridView={t("grid")} listView={t("list")} />
+            <GalleryMode />
+            <div role="tabpanel" aria-label={"games"} style={{paddingTop: "5px"}}>
+                { (viewMode === "GRID") ? <GamesGalleryGrid /> : <GamesGalleryList /> }
+            </div>
         </>
     )
 }
